Add image support for various arm archs
Traefik releases supported versions not only for armhf, but also for armv5, 6 and 7.
If the image is built on e.g. a raspberry pi 3, the step to determine supported arch fails and the image is not built.
With this change it is possible to build actual native images with support by official traefik binaries.